# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_post",
        "type": "object"
      },
...
      "DocumentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "effective_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "evidence_allowed": {
            "title": "Evidence Allowed",
            "type": "boolean"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "policy",
              "report",
              "record",
              "contract",
              "plan",
              "questionnaire",
              "statement",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "line_count": {
            "title": "Line Count",
            "type": "integer"
          },
          "metadata_source": {
            "enum": [
              "rule",
              "model",
              "user"
            ],
            "title": "Metadata Source",
            "type": "string"
          },
          "scope": {
            "anyOf": [
              {
                "enum": [
                  "internal-systems",
                  "customer-product",
                  "production",
                  "employees",
                  "vendors-and-contractors"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "source": {
            "enum": [
              "sample",
              "upload",
              "drive",
              "statement"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "final",
              "draft"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "source",
          "kind",
          "status",
          "effective_date",
          "scope",
          "evidence_allowed",
          "metadata_source",
          "line_count",
          "created_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "DocumentPatch": {
        "additionalProperties": false,
        "description": "A visitor override (spec 5 step 3): only the fields sent change; metadata_source becomes 'user'.",
        "properties": {
          "effective_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "evidence_allowed": {
            "title": "Evidence Allowed",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "policy",
              "report",
              "record",
              "contract",
              "plan",
              "questionnaire",
              "statement",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "scope": {
            "anyOf": [
              {
                "enum": [
                  "internal-systems",
                  "customer-product",
                  "production",
                  "employees",
                  "vendors-and-contractors"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "status": {
            "enum": [
              "final",
              "draft"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "DocumentPatch",
        "type": "object"
      },
      "DocumentUpdated": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "effective_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "evidence_allowed": {
            "title": "Evidence Allowed",
            "type": "boolean"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "policy",
              "report",
              "record",
              "contract",
              "plan",
              "questionnaire",
              "statement",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "line_count": {
            "title": "Line Count",
            "type": "integer"
          },
          "metadata_source": {
            "enum": [
              "rule",
              "model",
              "user"
            ],
            "title": "Metadata Source",
            "type": "string"
          },
          "redecided": {
            "title": "Redecided",
            "type": "integer"
          },
          "scope": {
            "anyOf": [
              {
                "enum": [
                  "internal-systems",
                  "customer-product",
                  "production",
                  "employees",
                  "vendors-and-contractors"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "source": {
            "enum": [
              "sample",
              "upload",
              "drive",
              "statement"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "final",
              "draft"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "source",
          "kind",
          "status",
          "effective_date",
          "scope",
          "evidence_allowed",
          "metadata_source",
          "line_count",
          "created_at",
          "redecided"
        ],
        "title": "DocumentUpdated",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineOut": {
        "properties": {
          "n": {
            "title": "N",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "n",
          "text"
        ],
        "title": "LineOut",
        "type": "object"
      },
      "LinesOut": {
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineOut"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "document_id",
          "filename",
          "lines"
        ],
        "title": "LinesOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VART",
    "version": "2.0.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "description": "Every document in the workspace, sample, uploaded and statements, oldest first.",
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Api Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "Multipart upload of one file. Parsed in memory, redacted, classified, chunked; the bytes are not\nstored. 422 with a sentence for a refused file; 429 per network (`upload`, 60 an hour); 503 when full.\nA body over Vercel's 4.5 MB limit gets the platform's own 413 (not JSON) before the app sees it.",
        "operationId": "upload_document_api_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorOut"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "Refused input: a sentence, or FastAPI's validation list"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/sample": {
      "post": {
        "description": "The sample company's documents (not redacted, not counted against the upload limit). Idempotent.",
        "operationId": "load_sample_documents_api_documents_sample_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response Load Sample Documents Api Documents Sample Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Load Sample Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "description": "409 when a run used the document (reset the workspace to start over).",
        "operationId": "delete_document_api_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Delete Document",
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "description": "Override metadata; every answer that used this document is decided again with no model call.",
        "operationId": "update_document_api_documents__document_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUpdated"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Update Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{document_id}/lines": {
      "get": {
        "description": "Stored (redacted) lines `from`..`to`, at most 200.",
        "operationId": "document_lines_api_documents__document_id__lines_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "From",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorOut"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "Refused input: a sentence, or FastAPI's validation list"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Document Lines",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}